chore: TDET-50 Moved around some folder hierarchies
